requirements.txt
numpy==1.21.6
opencv_python==4.5.4.58
opendr==0.78
pointnet2_ops==3.0.0
PyYAML==6.0
PyYAML==6.0.1
scikit_image==0.19.2
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tqdm==4.64.0
wandb==0.12.15
